% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_methods.R
\name{adjust.bl}
\alias{adjust.bl}
\title{Adjust a bl object with observed data}
\usage{
\method{adjust}{bl}(bl_obj, D)
}
\arguments{
\item{bl_obj}{A bl object}

\item{D}{A matrix of observed data}
}
\value{
Returns a bl_adjust object
}
\description{
Adjust a bl object with observed data
}
