% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_adjust_methods.R
\name{adjust.bl_adjust}
\alias{adjust.bl_adjust}
\title{Adjust a bl object with observed data}
\usage{
\method{adjust}{bl_adjust}(bl_obj, D, sequential = TRUE)
}
\arguments{
\item{bl_obj}{A bl_adjust object}

\item{D}{A matrix of observed data}

\item{sequential}{Update beliefs sequentially using existing data?}
}
\value{
Returns a bl_adjust object
}
\description{
Adjust a bl object with observed data
}
