% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_methods.R
\name{adjust.bs}
\alias{adjust.bs}
\title{Adjust a belief structure with observed data}
\usage{
\method{adjust}{bs}(bs_obj, D)
}
\arguments{
\item{bs_obj}{A belief structure}

\item{D}{A matrix of observed data}
}
\value{
Returns a adjusted belief structure (adj_bs) object
}
\description{
Adjust a belief structure with observed data
}
