% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_methods.R
\name{adjust}
\alias{adjust}
\title{Adjust a belief structure with observed data}
\usage{
adjust(obj, d, ...)
}
\arguments{
\item{obj}{A belief structure calculated from \code{belief_structure()}.}

\item{d}{A numerical vecotr of 1D matrix of observed data.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns a adjusted belief structure (\code{adj_belief_structure})
object
}
\description{
Adjust a belief structure with observed data
}
