% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_constrained.R
\name{adjust_constrained.bl}
\alias{adjust_constrained.bl}
\title{Adjust a bl object with observed data}
\usage{
\method{adjust_constrained}{bl}(bl_obj, D, constraint = NULL)
}
\arguments{
\item{bl_obj}{A bl_adjust object}

\item{D}{A matrix of observed data}

\item{constraint}{A list of constraints to be passed to CVXR}
}
\value{
Returns a bl_adjust object
}
\description{
Adjust a bl object with observed data
}
