% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_methods.R
\name{belief_structure}
\alias{belief_structure}
\title{Create belief structure object}
\usage{
belief_structure(exp_x, exp_d, cov_xd, var_x, var_d)
}
\arguments{
\item{exp_x}{Prior expectation of X. Expects either a numeric vector or 1D
matrix.}

\item{exp_d}{Prior expectation of D. Expects either a numeric vector or 1D
matrix.}

\item{cov_xd}{Prior covariance of X and D}

\item{var_x}{Prior variance of X}

\item{var_d}{Prior variance of D}
}
\value{
Returns a belief structure (\code{belief_structure}) object
}
\description{
Create belief structure object
}
