% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{bl}
\alias{bl}
\title{Crate a Bayes linear object}
\usage{
bl(exp_X, exp_D, cov_XD, var_X, var_D)
}
\arguments{
\item{exp_X}{Prior expectation of X}

\item{exp_D}{Prior expectation of D}

\item{cov_XD}{Prior covariance of X and D}

\item{var_X}{Prior variance of X}

\item{var_D}{Prior variance of D}
}
\value{
Returns a bl object
}
\description{
Crate a Bayes linear object
}
