% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_methods.R
\name{bs}
\alias{bs}
\title{Crate a belief structure}
\usage{
bs(E_X, E_D, cov_XD, var_X, var_D)
}
\arguments{
\item{E_X}{Prior expectation of X}

\item{E_D}{Prior expectation of D}

\item{cov_XD}{Prior covariance of X and D}

\item{var_X}{Prior variance of X}

\item{var_D}{Prior variance of D}
}
\value{
Returns a belief structure (bs) object
}
\description{
Crate a belief structure
}
