% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_ellipse}
\alias{calc_ellipse}
\title{Calculates uncertainty ellipse of a bivariate random variable}
\usage{
calc_ellipse(prob, mu, sigma)
}
\arguments{
\item{prob}{Sets ellipse width proportional to Normal}

\item{mu}{Ellipse centre}

\item{sigma}{Ellipse SD}
}
\value{
A tibble with x and y ellipse coordinates
}
\description{
Calculates uncertainty ellipse of a bivariate random variable
}
