% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_methods.R
\name{canonical}
\alias{canonical}
\title{Calculate the canonical directions and resolutions}
\usage{
canonical(obj, ...)
}
\arguments{
\item{obj}{Either a \code{belief_structure} or an \code{adj_belief_structure}
object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns a list of resolution matrix, canonical directions, and
canonical resolutions.
Note, the symmetric resolution matrix is used herein.
}
\description{
Calculate the canonical directions and resolutions
}
