% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_methods.R
\name{cantelli}
\alias{cantelli}
\title{Cantelli's inequality for a discrepancy S}
\usage{
cantelli(S)
}
\arguments{
\item{S}{Discrepancy term that's still to be named. Accepts either a numeric/vector/matrix object}
}
\value{
Returns a diagonal matrix with the Cantelli inequality calculations for each dimension on the diagonal
}
\description{
Cantelli's inequality for a discrepancy S
}
