% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_methods.R
\name{gbl_distance}
\alias{gbl_distance}
\title{GBL Distance}
\usage{
gbl_distance(gen_adj_exp, adj_exp, adj_var)
}
\arguments{
\item{Ec_adj}{CVXR Variable object}

\item{E_adj}{Standard BL adjusted expectation}

\item{var_adj}{Standard BL adjusted variance}
}
\value{
Returns the distance between \code{Ec_adj} and \code{E_adj} in the inner product space defined by \code{var_adj}
}
\description{
GBL Distance
}
