% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_methods.R
\name{gen_belief_structure}
\alias{gen_belief_structure}
\title{Create a Generalised Bayes Linear belief structure}
\usage{
gen_belief_structure(exp_x, exp_d, cov_xd, var_x, var_d, constraint)
}
\arguments{
\item{exp_x}{Prior expectation of X. Expects either a numeric vector or 1D
matrix.}

\item{exp_d}{Prior expectation of D. Expects either a numeric vector or 1D
matrix.}

\item{cov_xd}{Prior covariance of X and D}

\item{var_x}{Prior variance of X}

\item{var_d}{Prior variance of D}

\item{constraint}{Quoted constraint expression to be passed to \code{CVXR}.
See the README for examples on how to specify this.}
}
\value{
Returns a GBL belief structure (\code{gen_belief_structure}) object
}
\description{
Similar to a standard belief structure but with the ability to specify a
constrained solution space
}
