% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv.R
\name{inv}
\alias{inv}
\title{Calculates the Moore-Penrose generalized inverse of a matrix X.}
\usage{
inv(x)
}
\arguments{
\item{x}{Matrix for which the Moore-Penrose inverse is required.}
}
\value{
A MP generalized inverse matrix for X.
}
\description{
Calculates the Moore-Penrose generalized inverse of a matrix X.
}
