% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{one_dimension}
\alias{one_dimension}
\title{One Dimensional Example}
\format{
An object with class \code{bs}
}
\source{
Goldstein, Michael, and David Wooff. Bayes linear statistics: Theory and methods.
Vol. 716. John Wiley & Sons, 2007.
}
\usage{
one_dimension
}
\description{
A belief structure object of the one-dimensional problem from `Bayes Linear Statistics'.
}
\keyword{datasets}
