% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oral_glucose}
\alias{oral_glucose}
\title{Oral Glucose Tolerance Test}
\format{
An object with class \code{bs}
}
\source{
Goldstein, Michael, and David Wooff. Bayes linear statistics: Theory and methods.
Vol. 716. John Wiley & Sons, 2007.
}
\usage{
oral_glucose
}
\description{
A belief structure object of the oral glucose problem from `Bayes Linear Statistics'.
}
\keyword{datasets}
