% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_methods.R
\name{print.bs}
\alias{print.bs}
\title{Create a belief structure}
\usage{
\method{print}{bs}(x, ...)
}
\arguments{
\item{E_X}{Prior expectation of X}

\item{E_D}{Prior expectation of D}

\item{cov_XD}{Prior covariance of X and D}

\item{var_X}{Prior variance of X}

\item{var_D}{Prior variance of D}
}
\value{
Returns a belief structure (\code{bs}) object
}
\description{
Create a belief structure
}
