% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_methods.R
\name{resolution}
\alias{resolution}
\title{Calculates the resolution of an adjusted belief structure}
\usage{
resolution(obj, ...)
}
\arguments{
\item{obj}{An adjusted or standard belief structure object}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns a vector of calculated resolutions.
}
\description{
Calculates the resolution of an adjusted belief structure
}
