% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_methods.R
\name{solve_constrained_expectation}
\alias{solve_constrained_expectation}
\title{Solves for the constrained expectation of a GBL belief structure}
\usage{
solve_constrained_expectation(adj_exp, adj_var, constraint)
}
\arguments{
\item{constraint}{Quoted constraint expression to be passed to \code{CVXR}.
See the README for examples on how to specify this.}
}
\value{
The constrained adjusted expectation
}
\description{
Solves for the constrained expectation of a GBL belief structure
}
