% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Bayes linear update function}
\usage{
update(exp_X, exp_D, cov_XD, var_X, var_D, D)
}
\arguments{
\item{exp_X}{matrix}

\item{exp_D}{matrix}

\item{cov_XD}{matrix}

\item{var_X}{matrix}

\item{var_D}{matrix}

\item{D}{matrix}
}
\value{
list
}
\description{
Bayes linear update function
}
