% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bl_methods.R
\name{adjust}
\alias{adjust}
\title{Adjust either a belief structure or adjusted belief structure with observed data}
\usage{
adjust(obj, D, ...)
}
\arguments{
\item{obj}{A belief structure}

\item{D}{A matrix of observed data}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Returns a adjusted belief structure (\code{adj_bs}) object
}
\description{
Adjust either a belief structure or adjusted belief structure with observed data
}
