% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{bl}
\alias{bl}
\title{Crate a Bayes linear object}
\usage{
bl(exp_X, exp_D, cov_XD, var_X, var_D)
}
\arguments{
\item{exp_X}{matrix}

\item{exp_D}{matrix}

\item{cov_XD}{matrix}

\item{var_X}{matrix}

\item{var_D}{matrix}
}
\value{
Returns a bl object
}
\description{
Crate a Bayes linear object
}
