% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cantelli}
\alias{cantelli}
\title{Cantelli's inequality}
\usage{
cantelli(s)
}
\arguments{
\item{s}{constraint discrepancy vector. Expects a numeric vector.}
}
\value{
Returns a diagonal matrix with the inequality calculations on the
diagonals
}
\description{
Cantelli's inequality
}
