% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv}
\alias{inv}
\title{Calculates the Moore-Penrose generalized inverse of a matrix X.}
\usage{
inv(...)
}
\arguments{
\item{...}{Passes on arguments to MASS::ginv}
}
\value{
A MP generalized inverse matrix for X.
}
\description{
Calculates the Moore-Penrose generalized inverse of a matrix X.
}
