% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{one_dimension}
\alias{one_dimension}
\title{One-dimensional example}
\format{
A bl object of prior beliefs
}
\source{
Goldstein, Michael, and David Wooff. Bayes linear statistics: Theory and methods.
Vol. 716. John Wiley & Sons, 2007.
}
\usage{
one_dimension
}
\description{
Simple example as used in `Bayes Linear Statistics'
}
\keyword{datasets}
