% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{oral_glucose}
\alias{oral_glucose}
\title{Oral Glucose Tolerance Test}
\format{
A bl object of prior beliefs
}
\source{
Goldstein, Michael, and David Wooff. Bayes linear statistics: Theory and methods.
Vol. 716. John Wiley & Sons, 2007.
}
\usage{
oral_glucose
}
\description{
Oral glucose tolerance test used in `Bayes Linear Statistics'
}
\keyword{datasets}
