% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl_methods.R
\name{solve_constrained_expectation}
\alias{solve_constrained_expectation}
\title{Solves for the constrained expectation of a GBL belief structure}
\usage{
solve_constrained_expectation(E_adj, var_adj, constraint)
}
\arguments{
\item{E_adj}{Standard BL adjusted expectation}

\item{var_adj}{Standard BL adjusted variance}

\item{constraint}{Constraint expression. See the README for examples on how to specify this.
\code{bayeslinear} currently assumes that constraints are convex and solves using CVXR.}
}
\value{
The constrained adjusted expectation
}
\description{
Solves for the constrained expectation of a GBL belief structure
}
